libdir = $(CURDIR)/lib
Using a relative path to find shared libraries means when you call the
resulting executable, library lookup is performed relative to your
current location, so unless I'm calling it from the same directory in
which the build was performed, it will fail. Specifying an absolute path
fixes this issue. This might still not be the "right" way to handle
non-standard shared library locations but it's certainly an improvement
over what I had.